Update README.md
I did some changes to the How to Use? Part. I put up a reference to the link so if a user clicks the link it will direct them to the Google Web Store for the extension! I really think that it is necessary so! 
Great extension, Really loved it.
Thanks!